file system cleanup and reorganize
