--- util/listing_parser.py
@@ -1,3 +1,0 @@
-
-def parse_listing_entry(listing):
-    print('test')
